debug: fixing build steps

+++ bin/run_tests.sh
@@ -1,31 +1,22 @@
 #!/bin/bash
 
-set -e
-
-SELENIUM_LOG=$(mktemp /tmp/selenium.XXXXXXXX)
-METEOR_LOG=$(mktemp /tmp/meteor.startup.XXXXXXXX)
-
-if [[ $* == *--headless* ]]
-then
-  echo "Starting Selenium headless server..."
-  xvfb-run --server-args="-screen 0, 1366x768x24" node_modules/selenium-standalone/bin/start-selenium > $SELENIUM_LOG 2>&1 &
-else
+if [ -z "$CI" ]; then
   echo "Starting Selenium server..."
+  SELENIUM_LOG=$(mktemp /tmp/selenium.XXXXXXXX)
   node_modules/selenium-standalone/bin/start-selenium > $SELENIUM_LOG 2>&1 &
+
+  echo "Starting Meteor App with DB: mongodb://localhost:27017/parallels_test"
+  METEOR_LOG=$(mktemp /tmp/meteor.startup.XXXXXXXX)
+  cd meteor-app
+  MONGO_URL="mongodb://localhost:27017/parallels_test" meteor run --settings settings.json > $METEOR_LOG 2>&1 &
+  cd ..
+
+  # Wait for Meteor to finish booting
+  tail -f $METEOR_LOG | while read LOGLINE
+  do
+     [[ $LOGLINE == *"=> App running"* ]] && pkill -P $$ tail
+  done
 fi
-
-echo "Starting Meteor App with DB: mongodb://localhost:27017/parallels_test"
-cd meteor-app
-MONGO_URL="mongodb://localhost:27017/parallels_test" meteor run --settings settings.json &
-cd ..
-
-# Wait for Meteor to finish booting
-tail -f $METEOR_LOG | while read LOGLINE
-do
-  echo "Line is $LOGLINE"
-   [[ $LOGLINE == *"=> App running"* ]] && pkill -P $$ tail
-done
-cd ../
 
 # Run integration tests
 node node_modules/cucumber/bin/cucumber.js tests/features
